requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
async-generator==1.10
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
colorama==0.4.4
cryptography==36.0.1
h11==0.13.0
idna==3.3
iniconfig==1.1.1
more-itertools==8.12.0
outcome==1.1.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.7
pytest==7.0.1
selenium==3.14.0
sniffio==1.2.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.8
wcwidth==0.2.5
wsproto==1.0.0
Faker~=13.3.4